% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cm_age_coefficients}
\alias{cm_age_coefficients}
\title{Age coefficients
for a set of age group bounds age_bounds (including the lower and upper bound as well as intermediate bounds),
return a vector giving length(age_bounds) - 1 numerical entries between 0 and 1 inclusive, depending on how much
each age group lies within the range age_min to age_max.}
\usage{
cm_age_coefficients(age_min, age_max, age_bounds)
}
\description{
Age coefficients
for a set of age group bounds age_bounds (including the lower and upper bound as well as intermediate bounds),
return a vector giving length(age_bounds) - 1 numerical entries between 0 and 1 inclusive, depending on how much
each age group lies within the range age_min to age_max.
}
