% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cm_backend_mcmc_test}
\alias{cm_backend_mcmc_test}
\title{Run MCMC}
\usage{
cm_backend_mcmc_test(
  R_base_parameters,
  params_priors,
  seed,
  burn_in,
  iterations,
  n_threads,
  classic_gamma
)
}
\description{
Run MCMC
}
