% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cm_backend_optimize_test}
\alias{cm_backend_optimize_test}
\title{Run optimization}
\usage{
cm_backend_optimize_test(
  R_base_parameters,
  params_priors,
  maxeval,
  ftol_abs,
  seed,
  n_threads
)
}
\description{
Run optimization
}
