% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cm_backend_simulate_v2}
\alias{cm_backend_simulate_v2}
\title{Run simulation}
\usage{
cm_backend_simulate_v2(
  parameters,
  n_run = 1L,
  seed = 0L,
  n_threads = 1L,
  r_file_out = ""
)
}
\description{
Run simulation
}
