% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cm_base_parameters_SEI3R}
\alias{cm_base_parameters_SEI3R}
\title{Get default simulation parameters, SEI3R model}
\usage{
cm_base_parameters_SEI3R(n_groups = 1, pop = cm_base_pop_SEI3R(n_groups))
}
\description{
Get default simulation parameters, SEI3R model
}
