% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cm_build_pop_SEI3R}
\alias{cm_build_pop_SEI3R}
\title{Build parameters for a single location, SEI3R model}
\usage{
cm_build_pop_SEI3R(
  dem_location,
  mat_location = "guess",
  dE = NULL,
  dIp = NULL,
  dIs = NULL,
  dIa = NULL,
  dC = NULL,
  contact = NULL,
  imm0 = NULL,
  u = NULL,
  y = NULL,
  fIp = NULL,
  fIa = NULL,
  fIs = NULL,
  omega = NULL,
  rho = NULL,
  tau = NULL,
  v = NULL,
  ev = NULL,
  wn = NULL,
  wv = NULL,
  A = NULL,
  B = NULL,
  D = NULL,
  season_A = NULL,
  season_T = NULL,
  season_phi = NULL,
  seed_times = NULL,
  dist_seed_ages = NULL,
  observer = NULL,
  schedule = NULL
)
}
\description{
Build parameters for a single location, SEI3R model
}
