% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cm_delay_gamma}
\alias{cm_delay_gamma}
\title{construct a delay distribution following a gamma distribution with mean mu and shape parameter shape.}
\usage{
cm_delay_gamma(mu, shape, t_max, t_step)
}
\description{
construct a delay distribution following a gamma distribution with mean mu and shape parameter shape.
}
