% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cm_delay_skip}
\alias{cm_delay_skip}
\title{Delay skip
construct a delay distribution that effectively skips the compartment}
\usage{
cm_delay_skip(t_max, t_step)
}
\description{
Delay skip
construct a delay distribution that effectively skips the compartment
}
