% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cm_evaluate_distribution_v2}
\alias{cm_evaluate_distribution_v2}
\title{Evaluate a distribution}
\usage{
cm_evaluate_distribution_v2(dist_code, steps = 101L, xmin = 0, xmax = -1)
}
\arguments{
\item{dist_code}{Distribution code.}

\item{steps}{Number of steps.}

\item{xmin, xmax}{Range over which to evaluate distribution. Leave unspecified for full range of distribution.}
}
\description{
Evaluate a distribution
}
