% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cm_get_demographics}
\alias{cm_get_demographics}
\title{Get demographics for a given location, with error checking.}
\usage{
cm_get_demographics(dem_location, n_groups = NULL)
}
\description{
Get demographics for a given location, with error checking.
}
