% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cm_get_matrices}
\alias{cm_get_matrices}
\title{Get matrices for a given location, with error checking.}
\usage{
cm_get_matrices(mat_location, dem_location = NULL)
}
\description{
Get matrices for a given location, with error checking.
}
