% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cm_high_risk_prevalence}
\alias{cm_high_risk_prevalence}
\title{get prevalence of morbidities that increase risk of Covid per age-group
by default, returns prevalence of every included morbidity per age-group
to return overall prevalence, set aggregate = TRUE
to account for mutli-morbidities, set correlation of people with multiple morbidities multimorbidity_corr}
\usage{
cm_high_risk_prevalence(country, aggregate = FALSE, multimorbidity_corr = 0.8)
}
\description{
get prevalence of morbidities that increase risk of Covid per age-group
by default, returns prevalence of every included morbidity per age-group
to return overall prevalence, set aggregate = TRUE
to account for mutli-morbidities, set correlation of people with multiple morbidities multimorbidity_corr
}
