% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cm_interpolate_cos}
\alias{cm_interpolate_cos}
\title{Cosine interpolate}
\usage{
cm_interpolate_cos(x, x0, y0, x1, y1)
}
\description{
smoothly interpolate between points (x0, y0) and (x1, y1) using cosine interpolation.
for x < x0, returns y0; for x > x1, returns y1; for x0 < x < x1, returns the cosine interpolation between y0 and y1
}
