% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cm_make_population}
\alias{cm_make_population}
\title{look up pop_region and return vector with the number of people in different age bands (5 year size)}
\usage{
cm_make_population(pop_region, n_age_groups)
}
\description{
look up pop_region and return vector with the number of people in different age bands (5 year size)
}
