% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cm_mean_hdi}
\alias{cm_mean_hdi}
\title{Return the mean and estimated highest density interval from a set of samples x}
\usage{
cm_mean_hdi(x, credMass = 0.95)
}
\description{
Return the mean and estimated highest density interval from a set of samples x
}
