% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cm_parameters_SEI3R}
\alias{cm_parameters_SEI3R}
\title{Build parameters for one or several locations, SEI3R model}
\usage{
cm_parameters_SEI3R(
  dem_locations,
  mat_locations = "guess",
  date_start = "2020-03-01",
  date_end = "2021-03-01",
  deterministic = T,
  processes = NULL,
  dE = NULL,
  dIp = NULL,
  dIs = NULL,
  dIa = NULL,
  dC = NULL,
  contact = NULL,
  imm0 = NULL,
  u = NULL,
  y = NULL,
  fIp = NULL,
  fIa = NULL,
  fIs = NULL,
  omega = NULL,
  rho = NULL,
  tau = NULL,
  v = NULL,
  ev = NULL,
  wn = NULL,
  wv = NULL,
  A = NULL,
  B = NULL,
  D = NULL,
  season_A = NULL,
  season_T = NULL,
  season_phi = NULL,
  seed_times = NULL,
  dist_seed_ages = NULL,
  observer = NULL,
  schedule = NULL
)
}
\description{
Build parameters for one or several locations, SEI3R model
}
