% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cm_resample}
\alias{cm_resample}
\title{Safe sampling (unlike sample, works even if x is only one number)}
\usage{
cm_resample(x, ...)
}
\description{
Safe sampling (unlike sample, works even if x is only one number)
}
