% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{cm_sample_fit}
\alias{cm_sample_fit}
\title{sample fit
TODO -- in progress... interface likely to change.}
\usage{
cm_sample_fit(fit, n, time1 = NULL)
}
\description{
sample fit
TODO -- in progress... interface likely to change.
}
