% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{cm_sample_parameters}
\alias{cm_sample_parameters}
\title{sample parameters
TODO -- in progress... interface likely to change.}
\usage{
cm_sample_parameters(fit, n)
}
\description{
sample parameters
TODO -- in progress... interface likely to change.
}
