% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cm_split_matrices_ex_in}
\alias{cm_split_matrices_ex_in}
\title{Split matrices
ex_in: bounds separate into contacts *ex*clusively between lower groups, and *in*clusively between higher groups.}
\usage{
cm_split_matrices_ex_in(parameters, bounds)
}
\description{
Split matrices
ex_in: bounds separate into contacts *ex*clusively between lower groups, and *in*clusively between higher groups.
}
