% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cm_translate_parameters}
\alias{cm_translate_parameters}
\title{return translated parameters to work with the backend}
\usage{
cm_translate_parameters(p)
}
\description{
i.e. fix any times expressed in dates to be expressed in days since date0.
}
