% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidm.R
\docType{package}
\name{covidm}
\alias{covidm}
\title{covidm: SARS-CoV-2 dynamic transmission model.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

