% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{set_populations}
\alias{set_populations}
\title{set populations for a model}
\usage{
set_populations(model, locations, atlas = "wpp", admin = NULL, refyear = NULL)
}
\description{
set populations for a model
}
